IoUring: Disable support for RECVSEND_BUNDLE for now (#15104)
Motivation:

We did disable the testsuite for io_uring in the past as it did cause
the GH runner to die. After debugging more it seems to be caused by a
bug in our RECVSEND_BUNDLE support. Because of this we should just
disable the support for now (until we figured out what is the problem)
and enable the testsuite again.

Modifications:

- Disable RECVSEND_BUNDLE support
- Enable testsuite for io_uring again

Result:

Testsuite for io_uring is used again and support for RECVSEND_BUNDLE is
temporary disabled