IoUring: Re-enable RECVSEND_BUNDLE
Motivation:

We previously disabled the support for RECVSEND_BUNDLE as it did cause failures on our CI infrastructure. This change enables it again after fixing the bug that could let the buffer ring end up in an invalid state.

Modifications:

Correctly track when we need to full the ring again

Result:

Fix support for RECVSEND_BUNDLE