Add "unpadded" support to Base64 encoder/decoder (#15091)
Add "unpadded" support to Base64 encoder/decoder

Motivation:

This patch aims to improve the flexibility of the base64 encoder to
allow it to be used more broadly as a general purpose base64 encoder by
supporting output with no padding. While RFC 4648 suggests that base64
is always padded, Base64 URL does not require padding in all
circumstances, especially when length is known ahead of time. The
implementation in Netty assumes the length is always the size of the
buffer. In this case the decoder can produce correct output regardless
of whether padding is added for either Base64 or Base64 URL encoding.

When the encoder receives an unpadded Base64 URL string today it is
possible that one or two bytes are omitted from the resulting buffer. By
detecting if these bytes were not processed we can infer the necessary
padding and produce the full length decoded source text. Similarly on
the encoding side we can omit the padding if the user elects to do so.

Modifications:

Decode:
The `ByteProcessor` on the decode loop processes 4 bytes at a time,
calling `decode4to3` using a byte[] buffer of size 4. When the
`ByteProcessor` completes and there is no padding, the final call to
`decode4to3` will not have been made. We can infer based on the index
into the 4 byte buffer how much padding is necessary and we can make the
final call to `decode4to3` with appropriate padding added.

Encode:
If no padding is necessary, we can simply omit the padding bytes at the
end of our processing rounds instead of appending them.

Result:

The base64 encoder will support encoding and decoding Base64 without
padding. The decoder will handle both modes gracefully, and the encoder
will by default add padding, as is suggested by the RFC, but will allow
the user to elect not to in the event they want padding omitted.

Co-authored-by: Norman Maurer <norman_maurer@apple.com>