Fix concurrent ByteBuf write access bug in adaptive allocator (#15131) (#15132)
Motivation:
The bulk-write methods ByteBuf.copy and setBytes on buffers from the
adaptive allocator were delegating to their root parents, which is
effectively a shared ByteBuf for the chunk. This can cause concurrent
write access to the chunk buffer, which is unsafe.

Modification:
Change the method implementations in AdaptiveByteBuf to no longer
delegate to the root parent implementation, but instead obtain the
underlying NIO ByteBuffer slices and operate on those in a way that
avoids concurrent manipulation of the position and limits of the root
parent underlying buffer.

Result:
No more data-races in the bulk-write setBytes and copy methods for
AdaptiveByteBuf.