Always correctly calculate the memory address of the ByteBuf even if sun.misc.Unsafe is not usable
Motivation:

We might not be able to use sun.misc.Unsafe and so might need to use JNI to obtain the memory address of the ByteBuf by using the internal ByteBuffer. When doing so we need to take special care when the ByteBuffer was obtained via internalNioBuffer(...) as this method
might return a ByteBuffer that is shared across different ByteBuf instances. Because of this we need to also take the position of the returned ByteBuffer and add it to the memory address that is returned via JNI. Failing to do so might result in data-corruption.

Modifications:

Correct calculate the memory address of the ByteBuf even if sun.misc.Unsafe is not usable.

Result:

Not more data-corruptions when sun.misc.Unsafe is not usable and our native SSL implementation is used or io_uring is used. Also fixes https://github.com/netty/netty/issues/15037