Fix signing keys with ML-DSA in CertificateBuilder
Motivation:
Because in Java 8 we can't talk about the `AssymetricKey` class, or otherwise get the parameters of keys, we relied on reflection to obtain the `getParams` method.
However, we did so using the concrete key type class, which may be an implementation specific class that is not visible to us.
This caused illegal access problems.

Modification:
Instead explicitly obtain the `getParams` method from the `AssymetricKey` interface, which is the public API that defines it.
This is safe because the `PublicKey` interface, in the relevant Java versions, extend the `AssymetricKey` interface.

Result:
We can now sign keys and certificates using ML-DSA keys, and the CertificateBuilderTest passes.