Extract Bootstrap and ServerBootstrap setup to a common method (#15134)
This commit refactors the `FixedChannelPoolTest` class to extract the
repetitive setup of Bootstrap and ServerBootstrap into a private
`bootstrap()` method. This improves code readability and reduces
redundancy across multiple test methods.

**Motivation:**

The `FixedChannelPoolTest` class contained repetitive code for setting
up `Bootstrap` and `ServerBootstrap` instances in multiple test methods.

This will also reduce the effort to create new tests that are coming as
part of
* https://github.com/netty/netty/pull/14723
* https://github.com/netty/netty/discussions/15133

**Modification:**

A new private method `bootstrap()` was introduced within the
`FixedChannelPoolTest` class. This method encapsulates the creation and
configuration of a `LocalAddress`, a `Bootstrap` configured for
`LocalChannel`, and a `ServerBootstrap` configured for
`LocalServerChannel`.