Optimize ByteBuf.setCharSequence for adaptive allocator (#15165) (#15170)
Motivation:
The `setCharSequence` method is used a fair bit when encoding HTTP
headers. I noticed the AdaptiveByteBuf was sending the call down a path
with a more indirections than strictly necessary.

Modification:
Implement `setCharSequence` in the `AdaptiveByteBuf` so we skip a level
of indirections. This is especially profitable for direct/off-heap
buffers, where we see up to a 50% performance improvement in the
benchmark. The benefit is greater the longer the string is.

Interestingly, the benchmark also show that it isn't profitable to do
this for the `getCharSequence` sibling method.

Result:
Direct buffers from the adaptive allocator now have much faster
`setCharSequence`, which will help with HTTP header encoding.

Co-authored-by: Chris Vest <christianvest_hansen@apple.com>