Include failure logs in cancelled builds
Motivation:
Sometimes a build gets stuck, and times out after 6 hours or so.
Effectively, the build enters the cancelled state.
We should capture surefire logs when this happens so we can debug.

Modification:
Capture and upload the surefire-reports directory when a PR build gets cancelled.

Result:
Easier to tell what a build gets stuck on.