Provide a protected constructor for UnpooledUnsafeDirectByteBuf with slice flag
Motivation:
Reduce memory allocation by avoiding slicing of a direct byte buffer if it is not needed.
Particularly, Cassandra server has its own pooling mechanism to manage direct buffers and to control the lifecycle of such buffers when they are sent to network layer Cassandra inherits a wrapper class from UnpooledUnsafeDirectByteBuf (org.apache.cassandra.net.BufferPoolAllocator.Wrapped). Currently, such wrapping cannot be done without an extra slicing of a direct byte buffer for every request. The existing slice option could help to eliminate the overhead, but it is provided only in a package private constructor, so it cannot be used by subclasses.

Modifications:
Add a new protected constructor for UnpooledUnsafeDirectByteBuf with slice flag, flag is added not as the last parameter because a constructor with another boolean flag at the end already exists.