Ensure OpenSsl.availableJavaCipherSuites does not contain null value
Motivation:
The way BoringSSL-related TLS 1.3 certificates are regisgtered and
then converted to their java equivalent name, will cause a null
value to be added to the set of OpenSsl.availableJavaCipherSuites.

Modifications:
While the underlying root cause (the differently named ciphers) is
not resolved with this changeset, the code now makes sure that
if a conversion has not been successful (and returning null as a
result), it is not added to the list of available cipher suites
and therefore acts as a safeguard to the caller.

A related, outdated javadoc has been fixed as well and regression
tests added.

Fixes #15214