Make JdkZlibEncoder accept Deflater.DEFAULT_COMPRESSION as level
Motivation:
Right now the default compression level inside the used Deflater
is defined as "-1", which according to zlib manual also defaults
to "6" which is used inside netty as the default. That said,
we should also allow to provide Deflater.DEFAULT_COMPRESSION (-1)
as an argument.

Modifications:
The range check has been modified to allow for -1 and it now uses
the constants instead of magic numbers. Small regression test
added which just makes sure that no exception is thrown when provided.

Result:
Fixes #15212