Add support for creating direct ByteBuffers from MemorySegments
Motivation:
Netty's performance, particularly in very busy systems, relies heavily on the ability to immediately free native memory when it's no longer used.

Since time immemorial, we have been relying on `sun.misc.Unsafe`, and various hacks based upon it, for quickly releasing/cleaning direct ByteBuffers.

From Java 24 onwards, any use of `Unsafe` has started to produce warnings, and future versions will deny access entirely.
For that reason, we no longer rely on `Unsafe` by default (it can still be explicitly enabled) when running on Java 24 or greater.

This creates a large performance gap that needs to be closed.
Fortunately, Java 22 introduced the `MemorySegment` and related (FFM) APIs, and it is possible to immediately release native memory using these APIs.

As Netty 4.2 is now also based on Java 8, we have access to `MethodHandles`, which allows to craft efficient mechanisms for calling the FFM APIs, without requiring them to be available at compile time.

Modification:

1.
The lifetime of a memory segment is tracked by the associated `Arena` instance, but it is not possible to obtain the `Arena` from a `MemorySegment`, nor from any `ByteBuffer` created from the `MemorySegment`.
Therefor, we need to track the lifetime of both, together.
The `CleanableDirectBuffer` interface is introduced for this purpose, and encapsulates both a `ByteBuffer` and its preferred mechanism for freeing it.

2.
The _Netty_ `Cleaner` API is then extended with a method for allocating direct buffers in the form of the `CleanableDirectBuffer`.
Putting this in the cleaner allow us to switch the buffer allocation and releasing behaviour depending on the Java version, as we already have version specific implementations.

3.
A new `CleanerJava24` implementation is added, which – as mentioned – uses method handles to create and arena, allocate a memory segment and a byte buffer from it, and then encapsulate both in a `CleanableDirectBuffer` instance.

4.
A new `PlatformDependent.allocateDirect(int)` method is added, which returns `CleanableDirectBuffer` instances depending on what `Cleaner` implementation we end up using.

5.
A so-called `Cleaner` implementation is also added for the cases where we previously used `PlatformDependent.allocateDirectNoCleaner(int)`.
This means that the optimal mechanism for allocating and releasing buffers are now abstracted behind a single common API.
This also allows us to remove all the hacks and conditional code-paths that relied on `PlatformDependent.hasDirectBufferNoCleanerConstructor()`.

6.
Replace all instances where we allocate direct buffers, including in our native transports, with the new `CleanableDirectBuffer` mechanism.

7.
Finally, deprecate all the APIs that bypass the `CleanableDirectBuffer` API.
We cannot remove these APIs, however, as many of them are public, and it would break compatibility.

Result:
We can once again have explicit control over direct buffer lifetimes on Java 24 and greater.

Furthermore, since the FFM APIs are stable, and aligned with the long-term direction of the Java platform, we likely won't have to worry about this functionality ever breaking again.