Compatibility fix for EpollSocketTcpMd5Test (#15273)
Motivation:
Not all Linux kernels or distros allow or support TCP_MD5SIG. The
EpollSocketTcpMd5Test should ignore errors stemming from this, and abort
itself instead of failing.

Modification:
Catch the channel exception and abort the EpollSocketTcpMd5Test if we
get and error with "Protocol not availeble" when we try to enable
TCP_MD5SIG.

Result:
These tests now abort instead of failing, when TCP_MD5SIG is not
supported.