Generate netty-bom that does not inherit from netty-parent (#15265)
Motivation:

We need to ensure our netty-bom does not have netty-parent as a parent
as otherwise it will also pull in the dependency management of the
parent which might lead to unexpected behaviour. Beside this we also
should ensure our native jars are marked as runtime dependency.

Modifications:

- Use plugin to generate the correct bom file and so not depend on the
parent
- Use correct runtime scope for native dependencies
- Fix netty-all generation to directly specify dependency versions and
not depend on the bom as this does not work (due the usage of the
plugin).
- Remove duplicated tcnative.version declaration as this is not needed
anymore (the plugin does the correct resulution).
- Remove usage of script during build as its not needed anymore

Result:

Correct dependency management