Kqueue: Delay removal from registration map to fix noisy warnings
Motivation:

In Kqueue it sometimes happened that we did see a WARN that the registration for a given id could not be found. The reason for this was that it is possible that the registration was already be cancelled by the user and so not present anymore.
In this case the logs are harmless but still confusing. We should better ensure we handle it better.

Modifications:

When the user cancels a registration just mark it as cancelled and remove it from the map one all events are processed.

Result:

Less confusing logging and more correct handling of cancelled registrations