Share the maven cache across builds
Motivation:
The local maven cache itself is OS agnostic, and can be shared across our build matrices to reduce the number of caches we keep.
The cache action itself, however, has differences between Windows and other operating systems, so we do need to distinguish Windows in the cache keys.

Modification:
Change all the cache actions to use keys that are only based on the hash of the pom files, i.e. to exclude other information about the specific build.

Except on Windows, where we include "windows" in the cache key name, so the actions/cache don't need to worry about OS specifics in file paths and archive formats.

Result:
Greater reuse of our maven caches in our builds.