Use checkAvailable(); in #readByte(); method (#15305)
Motivation:

Use checkAvailable(); in #readByte(); method for a better error message and a better consistency between all #read[...](); methods

Modifications:

#readByte(); method now uses checkAvailable(); method to check if there is a least 1 byte in the buffer.
If there is no bytes to read, there will always be an EOFException error but with the message :  "fieldSize is too long! Length is 1, but maximum is 0" (similar to other methods).

Result:
In case of error, an EOF exception with a message.