Make ByteBuf.memoryAddress work for MemorySegment buffers
Motivation:
To take full advantage of direct buffers, we need to be able to get the address of their native memory.
We have been relying on Unsafe for this, but now that we are starting to allocate native memory using MemorySegments, we have other and more portable options available.

Modification:
Make the CleanerJava24 extract the native memory address from the MemorySegment instances it allocates, and pass it to its CleanableDirectBuffer instances.

Add methods to the CleanableDirectBuffer interface for querying the native memory address.

Make PooledDirectByteBuf and PooledHeapByteBuf take advantage of these new methods on the CleanableDirectBuffer. Note: the Unsafe variants of these types will continue to use Unsafe as before.

Fix numerous cases where we assume that ByteBuf.hasMemoryAddress returning 'true' means Unsafe is available.
This is no longer the case, and we have to query for Unsafe separately.

Result:
We can now get the memory address of direct ByteBuf instances when using memory segments on Java 24+.

Fixes https://github.com/netty/netty/issues/15309