AbstractDnsRecord: equals() and hashCode() to ignore name field's case (#15328)
Motivation:

In certain situations, it was found that DNS server implementations may
return a response to a DNS question with a case other than that in the
original request. In these cases, netty was found to continue waiting
for additional responses, eventually timing out as if it had not
received an answer to the DNS question.

This PR intends to make DNS answers match DNS questions when they vary
only by case.

Modification:

AbstractDnsRecord's equals() and hashCode() methods now behave the same
regardless of the name field's case

Result:

Fixes #15327 on the `main` branch.