Fix bug in capacity(int) for the adaptive allocator
Motivation:
Calling `ByteBuf.capacity(int)` may cause internal allocation as the buffer is expanded.
This allocation can fail, and we should not leave the buffer in an inconsistent state if that happens.

Modification:
Restore the temporarily overwritten capacity if the internal allocation fails.

Result:
It's no longer possible to end up with zero-capacity buffers if internal allocation fails.