Testsuite: Ensure counter is updated before write happens so lastIdx … (#15374)
…is always correct

Motivation:

We should increment the counter (which is used for the lastIdx) before
we trigger another operation on the Channel. This is needed as any
operation might result in having `fireChannelRead(...)` called at the
end and so we could end up re-entrance into the `channelRead` method and
then use the wrong value for the lastIdx.

Modifications:

Move counter update before channel.write(...)

Result:

More robust tests