direct buffer preference should not depend on unsafe presence (#15381)
Motivation:
Currently, the preference if a direct buffer should be allocated by, for
example, `ByteBufAllocator.buffer()` or `ByteBufAllocator.ioBuffer()`
depends on the fact if unsafe is available. It seems like this was
originally introduced as only with unsafe it was possible to get the
cleaner of a ByteBuffer to clean it. However, with the changes
introduced in #15231, it doesn't make sense anymore to depend on this as
unsafe could be unavailable but MemorySegments will still be freeable.

Modification:
Switch the check for unsafe to a check that validates if the current
cleaner implementation is able to free the allocated buffers (each
implementation can do this except the no-op one). This ensures that
direct buffers are actually preferred when they're backed by something
cleanable and still unpreferred when unsafe is unavailable and buffers
cannot be freed in a controlled way.

Result:
Direct buffers will be preferred when they're cleanable in a controlled
way unrelated to the fact if unsafe is available.