Replace LinkedBlockingQueue with MPSCQueue in SingleThreadIoEventLoop (#15372)
Motivation:

Netty previously used MpscQueue as the taskQueue implementation in
EpollEventLoop, KQueueEventLoop, and NioEventLoop. However, these are
now deprecated.
To ensure consistent performance, I override
SingleThreadEventLoop.newTaskQueue to return an MpscQueue

This also helps avoid potential deadlocks when the EventLoop is used as
a virtual thread scheduler.

For example, 
thread A calls submit() and acquires the putLock of a blocking queue.
Meanwhile, the EventLoop-vt also calls submit(), but gets blocked trying
to acquire the same lock and parks. The EventLoop itself later calls
submit() and also parks. When thread A releases the putLock, it attempts
to wake up the EventLoop-vt, but the EventLoop is already parked,leading
to a deadlock.

Modification:

Replace LinkedBlockingQueue with MPSCQueue in SingleThreadIoEventLoop

Result:

---------

Co-authored-by: Norman Maurer <norman_maurer@apple.com>