Adaptive: Fix possible buffer corruption caused by incorrect setCharS… (#15391)
…equence(...) implementation

Motivation:


https://github.com/netty/netty/commit/31c033a1d95eb4a74002c5f17ea01c967b07add9
introduced a regression that can lead to data corruption when calling
setCharSequence(...) with a Charset that needs more bytes to encode the
sequence than CharSequence.length(). The problem is that we just
delegate to rootParent() without checking if the AdapativeByteBuf has
enough capacity itself. If this is not the case we might still end up to
write the CharSequence into the rootParent() and so override the content
of another ByteBuf that also uses the same parent.

Modifications:

- Correctly implement setCharSequence and also ensure writeCharSequence
is consistent
- Add unit tests

Result:

Fix regression introduced by
https://github.com/netty/netty/commit/31c033a1d95eb4a74002c5f17ea01c967b07add9