Improve JFR availability check Motivation:
Using the javaVersion to determine the default value for `io.netty.jfr.enabled` was not ideal. It did not work on JDK 8 versions with backported JFR, and it did not work on certain GraalVM versions without JFR support.

Modification:

Instead of the javaVersion, call `FlightRecorder.isAvailable()`, as discussed in #15397.

Result:

More reliable JFR detection.

Fixes #15382
Fixes #15397

(both untested unfortunately, I don't know exactly which versions to test with)