Move the SDKMAN init to the version check step in the Java EA version check workflow
Motivation:
We need the `sdk` executable on the PATH in the script, and each workflow step runs in a new shell context.

Modification:
Move sourcing the SDKMAN init script to the version check step, so it runs in the shell context where we need it.

Result:
The workflow should now be able to actually perform the version check.