Fix concurrent ByteBuffer access issue in AdaptiveByteBuf.getBytes (#… (#15410)
…15402)

Motivation:
The ByteBuf.getBytes methods should be thread-safe, as it is in theory
non-mutating. However, their implementations access the internal NIO
buffer and mutate the position and limit. The buffer is duplicated to
try and counter this, but the duplication is done after the position and
limit are changed on the potentially shared buffer.

Modification:
Make these methods duplicate the shared buffer first, then mutate the
position and limit.

Result:
No more spooky position and limits on these buffers when
AdaptiveByteBuf.getBytes is called concurrently on the same buffer by
multiple threads.

Co-authored-by: Chris Vest <christianvest_hansen@apple.com>