Allow size-classed chunks to be reused before they become empty
Size-classed chunks do not use pointer-bump allocation but instead has a freelist, so we can allocate from them as soon as they have any free space.
To encourage this, we can add them to the shared queue as soon as a magazine is done with them.
By the time they are plucked from the queue, they may have space again, and if they do not, they will simply return to the back of the queue again (or get actually released).