Introduce size-classes for the adaptive allocator (#15399)
Motivation:
The histograms in the adaptive allocator are best at looking for a
single, good chunk size.
However, in practice, we have a lot of sizes that follow various
patterns.
The histograms then tend to bias toward large sizes, which is wasteful
in space.
It also turns out to be wasteful in time, in the sense that a small
allocation can stick around for a long time, and prevent a large chunk
from being reused when the rest of it has been deallocated.
As such, the histogram-based magazines works best for fewer, larger
allocations, and we need something else for small and plentiful
allocations.

Modification:
- Refactor the internals to detangle the magazines from the histogram
implementation.
- Introduce abstractions so the choice of recording and choosing buffer
sizes, and the chunk allocation, is moved out of the magazines and into
a chunk controller.
- Move the allocation statistics into a histogram chunk controller, and
introduce a separate size-classed chunk controller.
- The size-classes are chosen from the observation that 1) most
allocations aim to be power-of-2 sized, but 2) a lot of such allocations
end up in framing protocols that add headers, checksums, etc. so the
actual desired size is a power-of-2-plus-a-bit.
- Introduce magazine groups. The group holds the magazine array and
deals with contention by expanding the array. We will have a group per
size class, along size a histogram-based group for large allocations.
- Introduce a size-classed chunk which allocates out of a free-list
instead of pointer-bump allocation. This greatly enhances the space
reuse of such chunks and allows them to absorb a lot of the memory
fragmentation that small allocations are prone to.
- Add an MpscIntQueue class that we use for said free-list. This is
based on the implementations found in JCTools.
- Fix a concurrent-access bug in AdaptiveByteBuf.getBytes methods, where
the internalNioBuffer had position and limit concurrently modified.

Result:
This should greatly reduce memory usage caused by chunks that hang
around too long due to fragmentation.

Fixes https://github.com/netty/netty/issues/15311