Fix chunk reuse bug in adaptive allocator
Motivation:
In the original PR adding size classes to the adaptive allocator, I missed a few places where we release chunks in the allocation path.
Chunks now need to be released from a magazine using the new `releaseFromMagazine` method, which for the size-classed chunks will attempt to return the chunks to the shared queue immediately.

Modification:
Change the remaining `Chunk.release()` calls in the magazine allocation path to `Chunk.releaseFromMagazine()`, like they should've been to begin with.

Result:
This greatly enhances chunk reuse, especially so the more in-use ByteBufs we have at any one time.

In my local testing I see up to a 60% reduction in memory usage, depending on the use case.