Motivation:
We would like to use ManualIoEventLoop but the current decision when to block and when to not block when running the IO tasks does not fit our use-case entirely. It would be great if we could provide our own context implementation so we could decide when to block and for how long based application logic.

Modification:

Make ManualIoEventLoop::run(IoHandlerContext context, long runAllTasksTimeoutNanos) public

Result:

Users are able to provide custom rules for when to block and for how long