Add an allocation pattern simulator
Motivation:
It's useful to simulate allocation patterns in a more controlled environment, and record their memory usage performance.

Modification:
Add the `AllocationPatternSimulator` program, which simulates allocation patterns in workloads with various volumes of concurrently allocated buffers, and concurrent access level to the allocator.

Result:
We can now visualize the improvements we've done to the adaptive allocator in Netty 4.2.3.