Add JFR events for PooledByteBufAllocator
Motivation:
As we work to improve our allocators, it's useful to be able to capture recordings of the allocation activity.

Modification:
- Extract the JFR events out of the adaptive allocator, and into a separate package.
- Add JFR event calls to the pooled allocator, to record when buffers and chunks are allocated and freed.
- Add a backdoor mechanism for the events to obtain the memory address of a buffer, potentially even after it's been freed. This is needed to record the JFR events.

Result:
We can now get JFR events for both the adaptive and the pooled allocator.
The events contain a type field to signal which allocator implementation created the event.

Fixes https://github.com/netty/netty/issues/13423