Avoid duplicate size class declarations
Motivation:
We should not have multiple copies of the adaptive size class array, as it will be harder to maintain if we want to change it.

Modification:
Add a package-protected getter that returns a clone, which the test can use.
Also did a little code clean up to silence some warnings.

Result:
The AdaptivePoolingAllocatorTest is now always in sync with the real size class array.