IoUring: Add the ability to allocate the buffers for the buffering in (#15457)
bulk

Motivation:

Before we did fill in a new buffer as soon as soon as we used one. This
resulted in having buffer allocations spread out and so we could observe
multiple TLB misses when the kernel tried to use the buffer ring. To
reduce this we should better fill the whole ring at once and so allow
the user to do a large allocation for the whole ring and then slice out
the buffers as need while also still allow to do one allocation per
buffer.

Modifications:

- Introduce IoUringBufferRingAllocator.allocateBulk(...) and use it from
the IoUringBufferRing
- Add abstract base class for IoUringBufferRingAllocator and extend it
from our existing allocators. These now support large allocations or
normal allocations

Result:

Better performance and much more flexibility