Make sure JFR support works in a modular runtime (#15468)
Motivation:

The recent addition of JFR is not yet tested in a modular runtime.

Changes:

Add testing of JFR support in a modular runtime.

Add missing `jdk.jfr` declaration to `netty-common` `module-info`
descriptor.

Make `AllocateChunkEvent` public, so we can emit this event in the JPMS
test suite.

Note:

Testing JFR events requires the presence of `RecordingStream` that is
only available since JDK 14. The build is to skip the compilation of the
flight recorder test before JDK 21. Which means that testing JFR
requires to use JDK 21 or more.