IoUring: Simplify submission strategy (#15476)
Motivation:

At some point I added some code to submit as soon as a channel becomes
unwritable. This was always kind of a hack and we should better not do
it. With all the recent changes we see the expected performance even
without this.

Modifications:

- Remove the usage of CompletionBuffer as its not neeeded anymore (we
still keep the class for now just in case)
- Don't trigger a submit when the Channel becomes unwritable.

Result:

Cleanup