Move the MemorySegment based Buffer implementation into the netty-buffer module
Motivation:
As our version baseline is Java 22, we don't need to have this implementation stowed away in a separate module.
When Java 25 is released, we can also consider bumping our baseline and removing the ByteBuffer and Unsafe implementations.

Modification:
Move the netty-buffer-memory-segment files to other modules;
- the buffer implementation moves into the netty-buffer module
- the micro-benchmarks move to the netty-microbench module

Result:
Better code organization.