IoUring: Add support for IORING_SETUP_CQE32 (#15487)
Motivation:

To support some operations we need to support IORING_SETUP_CQE32 as we
need extra space to pass informations back as part of the completion
event.

Modifications:

- Add extra field to IoUringIoEvent to be able to represent extra data
that is provided when IORING_SETUP_CQE32 is used.
- Pass through the extra data when needed
- Add testcase

Result:

Be able to use IORING_SETUP_CQE32 which is also a requirement for
https://github.com/netty/netty/issues/15475

---------

Co-authored-by: Chris Vest <christianvest_hansen@apple.com>